requirements.txt
kucoin-universal-sdk==1.1.1
python-dotenv==1.0.1

# Dépendances PancakeSwap
web3==6.15.1
requests==2.32.3

# Dépendances Telegram
python-telegram-bot==20.8